Add workflow to auto-add issues to bounties platform project
Since we've separated the frontend from the backend, we must track 2 repos in our bounties platform project. The free version of github only allows the project to reference one repo for free. So we are implementing this workflow to have issues be auto-added to the bounties project board from a second repo.

This workflow uses a PAT that must be renewed 1/25/25.